requirements.txt
# Vault
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
cryptography==43.0.1
python-jose[cryptography]==3.3.0  # JWT tokens
passlib==1.7.4

# Agent
anthropic==0.34.2
paramiko==3.4.1               # SSH client

# Audit Dashboard
streamlit==1.38.0

# Shared
python-dotenv==1.0.1
httpx==0.27.2                 # Agent → Vault HTTP calls

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx                         # TestClient for FastAPI
